backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
google-auth-oauthlib
google-auth-httplib2
google-api-python-client